+++ src/main/groovy/life/qbic/datamodel/dtos/business/services/Sequencing.groovy
@@ -20,9 +20,25 @@ class Sequencing extends AtomicProduct {
    * @param unitPrice The price in € per unit
    * @param unit The product unit
    * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+   * @deprecated 2.5.0
    */
-
+  @Deprecated
   Sequencing(String name, String description, double unitPrice, ProductUnit unit, String runningNumber) {
     super(name, description, unitPrice, unit, new ProductId(ProductType.SEQUENCING.toString(), runningNumber))
+  }
+
+  /**
+   * Basic product constructor.
+   *
+   * Checks that all passed arguments are not null.
+   *
+   * @param name The name of the product.
+   * @param description The description of what the product is about.
+   * @param unitPrice The price in € per unit
+   * @param unit The product unit
+   * @param runningNumber Number used in conjunction with ProductType{@link life.qbic.datamodel.dtos.business.services.ProductType} to identify product
+   */
+  Sequencing(String name, String description, double unitPrice, ProductUnit unit, long runningNumber) {
+    super(name, description, unitPrice, unit, new ProductId.Builder(ProductType.SEQUENCING.toString(), runningNumber).build())
   }
 }