+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 ``life.qbic.datamodel.dtos.business.services.MetabolomicAnalysis``, ``life.qbic.datamodel.dtos.business.services.DataStorage``
 
 **Fixed**
+* Override ``equals()`` method for ``life.qbic.datamodel.dtos.business.OfferId` and ``life.qbic
+.datamodel.dtos.business.TomatoId``properly
 
 **Dependencies**
 
@@ -21,8 +23,6 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Constructor using the deprecated ``life.qbic.datamodel.dtos.business.ProductId`` constructor for ``life.qbic.datamodel.dtos.business.services.Sequencing``, ``life.qbic.datamodel.dtos.business.services.SecondaryAnalysis`,
 ``life.qbic.datamodel.dtos.business.services.ProteomicAnalysis``, ``life.qbic.datamodel.dtos.business.services.ProjectManagement``, ``life.qbic.datamodel.dtos.business.services.PrimaryAnalysis`,
 ``life.qbic.datamodel.dtos.business.services.MetabolomicAnalysis``, ``life.qbic.datamodel.dtos.business.services.DataStorage``
-* Override ``equals()`` method for ``life.qbic.datamodel.dtos.business.OfferId` and ``life.qbic
-.datamodel.dtos.business.TomatoId``properly
 
 
 2.4.0 (2021-03-18)