+++ CHANGELOG.rst
@@ -27,6 +27,8 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Override ``equals()`` method for ``life.qbic.datamodel.dtos.business.OfferId`` and
   ``life.qbic.datamodel.dtos.business.TomatoId`` properly
 
+* Overide ``equals()`` method for ``life.qbic.datamodel.dtos.business.ProductId` properly
+
 **Dependencies**
 
 **Deprecated**